Former-commit-id: c8cccae9e20541f13ccab494010e79c09e1aaac8 [formerly de3ce9d7082a2f99677c5d7b90e088f0c3b38c90]
Former-commit-id: bad4364d6bda9d3df15d194e021749566e35f252
Former-commit-id: 218aae69a9f1f06e7b3c823986b31d65784dd44c

+++ Utilities/Firebot/firebot_win.bat
@@ -1,12 +1,14 @@
- @echo off
+@echo off
 
 set fdsroot=%~f1
 set fdsbasename=%~n1
 
-set update=%2
-set altemail=%3
-set usematlab=%4
-set emailto=%5
+set clean=%2
+set update=%3
+set altemail=%4
+set usematlab=%5
+set installed=%6
+set emailto=%7
 
 if NOT exist %fdsroot% (
   echo ***Error: the repository %fdsroot% does not exist
@@ -14,8 +16,15 @@ if NOT exist %fdsroot% (
   exit /b 1
 )
 
+set CURDIR=%CD%
+
 echo.
-echo   FDS repo: %fdsroot%
+echo      FDS repo: %fdsroot%
+echo run directory: %CURDIR%
+if %clean% == 1 echo cleaning repo: yes
+if %clean% == 0 echo cleaning repo: no
+if %update% == 1 echo updating repo: yes
+if %update% == 0 echo updating repo: no
 echo.
 
 :: -------------------------------------------------------------
@@ -29,8 +38,6 @@ set OMP_NUM_THREADS=1
 :: -------------------------------------------------------------
 ::                         setup environment
 :: -------------------------------------------------------------
-
-set CURDIR=%CD%
 
 if not exist output mkdir output
 if not exist history mkdir history
@@ -93,8 +100,6 @@ if NOT "%emailto%" == "" (
 
 echo Stage 0 - Preliminaries
 
-:: check if compilers are present
-
 echo. > %errorlog%
 echo. > %warninglog%
 echo. > %stagestatus%
@@ -105,29 +110,45 @@ echo             found get_time
 call :GET_TIME TIME_beg
 call :GET_TIME PRELIM_beg
 
+:: looking for Fortran
 ifort 1> %scratchfile% 2>&1
 type %scratchfile% | find /i /c "not recognized" > %counta%
 set /p nothave_ifort=<%counta%
 set have_ifort=1
 if %nothave_ifort% == 1 (
-  echo "***Fatal error: Fortran compiler not present"
+  echo             Fortran not found
+  echo "           firebot aborted"
   echo "***Fatal error: Fortran compiler not present" > %errorlog%
-  echo "firebot run aborted"
   call :output_abort_message
   exit /b 1
 )
 echo             found Fortran
 
+:: if -installed option is used use installed smokeview
+::    otherwise look for C to build smokeview and
+::    use installed smokeview if C not found
+if %installed% == 1 goto else1
 icl 1> %scratchfile% 2>&1
 type %scratchfile% | find /i /c "not recognized" > %countb%
 set /p nothave_icc=<%countb%
 if %nothave_icc% == 1 (
   set have_icc=0
-  echo "***Warning: C/C++ compiler not found - using installed Smokeview to generate images"
+  echo             C compiler not found - looking for Smokeview
+  call :is_file_installed smokeview|| exit /b 1
+  set smokeview=smokeview
+  echo             found smokeview
 ) else (
-  echo             found C/C++
+  echo             found C
 )
+goto endif1
+:else1
+  set have_icc=0
+  call :is_file_installed smokeview|| exit /b 1
+  set smokeview=smokeview
+  echo             found smokeview
+:endif1
 
+:: looking  for email
 if NOT exist %emailexe% (
   echo ***Warning: email client not found.   
   echo             firebot messages will only be sent to the console.
@@ -135,22 +156,17 @@ if NOT exist %emailexe% (
   echo             found mailsend
 )
 
-call :is_file_installed pdflatex|| exit /b 1
-echo             found pdflatex
-
-call :is_file_installed grep|| exit /b 1
-echo             found grep
-
-call :is_file_installed sed|| exit /b 1
-echo             found sed
-
 call :is_file_installed cut|| exit /b 1
 echo             found cut
 
 call :is_file_installed git|| exit /b 1
 echo             found git
 
-::*** looking for matlab
+call :is_file_installed grep|| exit /b 1
+echo             found grep
+
+call :is_file_installed make|| exit /b 1
+echo             found make
 
 where matlab 2>&1 | find /i /c "Could not find" > %OUTDIR%\stage_count0a.txt
 set /p nothavematlab=<%OUTDIR%\stage_count0a.txt
@@ -162,27 +178,32 @@ if %nothavematlab% == 1 (
   echo             matlab not found - VV and User guides will not be built
 )
 
+call :is_file_installed pdflatex|| exit /b 1
+echo             found pdflatex
+
+call :is_file_installed sed|| exit /b 1
+echo             found sed
+
 echo. 1>> %OUTDIR%\stage0.txt 2>&1
 
 :: revert FDS/Smokeview repository
 
-if %update == 0 skip_update
-if "%fdsbasename%" == "FDS-SMVgitclean" (
-   echo             reverting %fdsbasename% repository
+if %clean% == 0 goto skip_clean1
+   echo             cleaning %fdsbasename% repository
    cd %fdsroot%
-   git clean -dxf 1> Nul 2>&1
+   git clean -dxf -e win32_local 1> Nul 2>&1
    git add . 1> Nul 2>&1
    git reset --hard HEAD 1> Nul 2>&1
-
-)
+:skip_clean1
 
 :: update FDS/Smokeview repository
 
+if %update% == 0 goto skip_update1
 echo             updating %fdsbasename% repository
 cd %fdsroot%
 git fetch origin
 git pull 1>> %OUTDIR%\stage0.txt 2>&1
-:skip_update
+:skip_update1
 
 cd %fdsroot%
 git describe --long --dirty > %revisionfilestring%
@@ -206,7 +227,7 @@ echo             parallel debug
 
 cd %fdsroot%\FDS_Compilation\mpi_intel_win_64_db
 erase *.obj *.mod *.exe *.pdb 1> Nul 2>&1
-make VPATH="../../FDS_Source" -f ..\makefile mpi_intel_win_64_db 1> %OUTDIR%\makefdsd.log 2>&1
+call make_fds bot 1> %OUTDIR%\makefdsd.log 2>&1
 call :does_file_exist fds_mpi_win_64_db.exe %OUTDIR%\makefdsd.log|| exit /b 1
 call :find_warnings "warning" %OUTDIR%\makefdsd.log "Stage 1b, FDS parallel debug compilation"
 
@@ -214,7 +235,7 @@ echo             parallel release
 
 cd %fdsroot%\FDS_Compilation\mpi_intel_win_64
 erase *.obj *.mod *.exe *.pdb 1> Nul 2>&1
-make VPATH="../../FDS_Source" -f ..\makefile mpi_intel_win_64  1> %OUTDIR%\makefdsr.log 2>&1
+call make_fds bot 1> %OUTDIR%\makefdsr.log 2>&1
 call :does_file_exist fds_mpi_win_64.exe %OUTDIR%\makefdsr.log|| exit /b 1
 call :find_warnings "warning" %OUTDIR%\makefdsr.log "Stage 1d, FDS parallel release compilation"
 
@@ -222,18 +243,20 @@ call :find_warnings "warning" %OUTDIR%\makefdsr.log "Stage 1d, FDS parallel rele
 ::                           stage 2
 :: -------------------------------------------------------------
 
+if %installed% == 1 goto skip_build_cstuff
+if %have_icc% == 0 goto skip_build_cstuff
 echo Stage 2 - Building Smokeview
 
 echo             libs
 
 cd %fdsroot%\SMV\Build\LIBS\lib_win_intel_64
-call makelibs2 1>> %OUTDIR%\stage2a.txt 2>&1
+call makelibs bot 1>> %OUTDIR%\stage2a.txt 2>&1
 
 echo             debug
 
 cd %fdsroot%\SMV\Build\intel_win_64
 erase *.obj *.mod *.exe smokeview_win_64_db.exe 1> Nul 2>&1
-make -f ..\Makefile intel_win_64_db 1> %OUTDIR%\makesmvd.log 2>&1
+call make_smv_db -r bot 1> %OUTDIR%\makesmvd.log 2>&1
 call :does_file_exist smokeview_win_64_db.exe %OUTDIR%\makesmvd.log|| exit /b 1
 call :find_warnings "warning" %OUTDIR%\makesmvd.log "Stage 2a, Smokeview debug compilation"
 
@@ -241,10 +264,12 @@ echo             release
 
 cd %fdsroot%\SMV\Build\intel_win_64
 erase *.obj *.mod smokeview_win_64.exe 1> Nul 2>&1
-make -f ..\Makefile intel_win_64 1> %OUTDIR%\makesmvr.log 2>&1
+call make_smv -r bot 1> %OUTDIR%\makesmvr.log 2>&1
 
 call :does_file_exist smokeview_win_64.exe %OUTDIR%\makesmvr.log|| aexit /b 1
 call :find_warnings "warning" %OUTDIR%\makesmvr.log "Stage 2b, Smokeview release compilation"
+set smokeview=%fdsroot%\SMV\Build\intel_win_64\smokeview_win_64.exe
+:skip_build_cstuff
 
 :: -------------------------------------------------------------
 ::                           stage 3
@@ -261,9 +286,9 @@ call :find_warnings "warning" %OUTDIR%\makefds2ascii.log "Stage 3, Building FDS/
 
 if %have_icc% == 1 (
   echo             background
-  cd %fdsroot%\Utilities\background\intel_win_32
+  cd %fdsroot%\Utilities\background\intel_win_64
   erase *.obj *.mod *.exe 1> Nul 2>&1
-  make -f ..\Makefile intel_win_32 1> %OUTDIR%\makebackground.log 2>&1
+  call make_background bot 1> %OUTDIR%\makebackground.log 2>&1
   call :does_file_exist background.exe %OUTDIR%\makebackground.log
   call :find_warnings "warning" %OUTDIR%\makebackground.log "Stage 3, Building FDS/Smokeview utilities"
 ) else (
@@ -284,14 +309,14 @@ echo             debug mode
 
 :: run cases
 
-cd %fdsroot%\Verification\
+cd %fdsroot%\Verification\scripts
 call Run_FDS_cases %debug% 1> %OUTDIR%\stage4a.txt 2>&1
 
 :: check cases
 
 set haveerrors_now=0
 echo. > %OUTDIR%\stage_error.txt
-cd %fdsroot%\Verification\
+cd %fdsroot%\Verification\scripts
 call Check_FDS_cases 
 
 :: report errors
@@ -303,13 +328,19 @@ echo             release mode
 :: run cases
 
 cd %fdsroot%\Verification\
+if %clean% == 0 goto skip_clean2
+   echo             cleaning Verification directory
+   git clean -dxf -e win32_local 1> Nul 2>&1
+:skip_clean2
+
+cd %fdsroot%\Verification\scripts
 call Run_FDS_cases %release% 1> %OUTDIR%\stage4b.txt 2>&1
 
 :: check cases
 
 set haveerrors_now=0
 echo. > %OUTDIR%\stage_error.txt
-cd %fdsroot%\Verification\
+cd %fdsroot%\Verification\scripts
 call Check_FDS_cases
 
 :: report errors
@@ -335,8 +366,8 @@ call :GET_TIME MAKEPICS_beg
 echo Stage 5 - Making pictures
 echo             FDS verification cases
 
-cd %fdsroot%\Verification\
-call MAKE_FDS_pictures 64 1> %OUTDIR%\stage5.txt 2>&1
+cd %fdsroot%\Verification\scripts
+call MAKE_FDS_pictures %smokeview% 1> %OUTDIR%\stage5.txt 2>&1
 
 if %have_matlab%==0 goto skip_matlabplots
 echo             matlab verification plots
